build(github): make `cafkafk` codeowner of everything
This makes me the codeowner of all files, meaning any PR will request me
for review. While this should (hopefully) be changed later to offload
the review burden, it's nice in the start of a project to ensure things
get prompt reviews.

Ideally, we'll have codeowners for specific sections of the code, e.g.
critical parts, as well as parts where people have domain specific
knowledge.

Signed-off-by: Christina Sørensen <christina@cafkafk.com>