bug(api): handle branch names with slashes in them
When dealing with branches, don't capture just one slice of the path as
the branch name, but every part of the path from that point on. This
lets us easily deal with branches that have slashes in their names.

Since the branch is always the last part of the URI, and the `/b/` or
`/branch/` part routes things here anyway, capturing the rest of the
path is safe. Only the path is captured, query strings are not, those
aren't part of the routing.

Fixes #19.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>