refactor(releases): Lift out the release list fetching
All of the forges Rime currently supports have a very similar way to
query the releases of a particular repository, the major difference - as
far as Rime is concerned - is the URL. This is clearly shown by all the
`forge_api_get_releases()` functions essentially being copy pastes of
each other.

To reduce the amount of copy paste, lift that shared code out into
`ForgeReleases::from_url()`, with the previous helper functions
remaining as thin wrappers around it.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>