feat(prereleases): Support including pre-releases
Built on top of the previous refactorings, this introduces the
`?include_prereleases=[true|false]` query option for the repo endpoints.
If pre-releases are to be considered, then Rime will return the latest
release, otherwise the latest one that is not a pre-release (assuming
the forge supports prereleases).

If the parameter is not specified, it defaults to `false`.

Fixes #11.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>