bug(cli): Gracefully handle partial configs
If the configuration file read only contains partial configuration,
gracefully fall back to default values for settings that were not
specified.

This is achieved by not accessing the settings directly, but through a
wrapper that can handle the necessary fallback. To enforce that, the
struct members are no longer public.

Fixes #26.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>