refactor(api): Lift out a lot of shared code
All of the forges were using *very* similar code, copy pastes of each
other with very little differences. To remedy the situation, this patch
fundamentally changes how the various forges are implemented: there is
now a common `Forge` trait, the forges implement *that* (and any
additional functionality they may require on top).

Default handlers and routing are provided, so the forges do not need to
implement any of those things. All the forges need to implement, are a
couple of functions that return Strings (or an error), the rest is
handled by common code.

If a forge does not implement a particular type of endpoint, the
appropriate function that'd construct the URL to retrieve the tarball
from there, should return `Err(ForgeError::EndpointUnavailable)`, and
the shared handlers will do the right thing, and turn that into a 404
with a descriptive error message.

The way the Forge objects are used, is that they get added as an
extension layer to the forge-specific routes, thus, the generic handlers
can pick them out, and call the forge-specific function implementations
they provide.

Fixes #29.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>