feat(semver): Add an endpoint for semver locking
This adds a new `/semver/:version` (and `/s/:version`) route to all
forges that implement release listing. The specified version is parsed
as a semver version requirement (more about that here[^1]), and releases
for the given repo are matched against it.

The matching is done by comparing the sanitized tag name: if the tag
begins with `v`, that character is stripped, and the rest is parsed as a
semantic version string; if the tag begins with the repo name and a
dash (another common pattern in tagging), then the `{repo}-` prefix is
removed, and the rest is parsed as a semantic version string; if neither
case applies, we try to parse the tag name as a semantic version as is.

If we can't parse the sanitized tag name as a semantic version, then we
consider it non-matching. Otherwise it is compared to the given version
requirement.

This allows many forms of version requirement specifications, but is
limited. Another way will be added later that allows specifying more
complex ranges in the query string.

 [^1]: https://doc.rust-lang.org/cargo/reference/specifying-dependencies.html

Fixes #10.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>