refactor(api): Turn auto-discovery into a "Forge"
There was an annoying issue with the previous implementation of
auto-discovery: it captured the *entire* path, even paths that would
otherwise go unrouted.

This new implementation is just an implementation of the Forge trait,
one that does some auto-detection, and then dispatches to the detected
forge - if any. The advantage of this is that the implementation is
simpler, and autodetection is only done for routes that would be routed.
Furthermore, Rime no longer does a redirect, but handles it all
internally.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>