docs(gitlab): Update the comment about url encoding
Other parts of the GitLab API use the user and repo names without URL
encoding, so there really is no need to encode *those*. Thus, the only
remaining part that needs encoding, is the `/` path separator, which we
manually url-encoded to `%2F`.

If we encoded the user and repo too, that might lead to surprising
results later down the line. To not hide problems, leaving them as-is is
the preferable way.

Update the comment on `Gitlab::get_api_releases_url` accordingly.

Fixes #12.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>