refactor(routes): Lift routing out of the Forge trait
There's no need to have the routing be part of the trait. It was only
there for the sake of convenience, and we can achieve similar levels of
convenience in ways that doesn't require routing in the trait itself.

Like simply lifting out the routing functions, and making them take an
argument that implements the Forge trait (+ some others so we can turn
it into a DynForge).

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>