refactor(forge): allow some shortcuts for implementors
Allow implementors of the Forge trait to shortcut the latest release and
semantic version endpoints. As in, if a Forge implements these features
itself, we do not need to re-implement it in Rime, but can dispatch the
request to the forge itself.

Default implementations are provided for both
`get_tarball_url_for_semantic_version()` and
`get_tarball_url_for_latest_release()`, both of which return `Ok(None)`.

The reason we're using `Result<Option<String>, ForgeError>` here, is
because not implementing these is not an error. Not implementing these
means the handler will do a fallback to using `get_api_releases_url()`
and `ForgeReleases`.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>