refactor(handlers): Lift out some shared code
Both `get_tarball_url_for_latest_release()` and
`get_tarball_url_for_semantic_version()` were using a very similar
pattern to fetch available releases, reduce it down to one, and
construct a redirect url if one was found, or return an error if not.
Both functions had a way to shortcut fetching and iterating over the
releases, if the forge supports it - that part was lifted out aswell!

This lifts that code out into a shared function, so it does not need to
be repeated in both functions.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>