feat(forge): try extracting a semver more aggressively
When trying to extract a semantic version from a tag name, do so more
aggressively: try parsing its slices, starting from the beginning, and
stop as soon as a semantic version is found. If none are found, return
an error.

This replaces the former special casing of the `v` and `{repo}-`
prefixes. While the new method is a little less efficient, it's not
terribly so, but it is far more flexible.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>