fix(forgejo): Revert "Improved auto-detection"
This reverts commit 41c5ec96229aa38e90f5f928eb09e3a416830343, because
the auto-detection done in it turned out to be unreliable. The old
method might be a tiny bit more expensive, not as nice, but it is
reliable, because it does not depend on the Forgejo instance
configuration.

Also re-enable the tests that were failing due to the naive
auto-detection breaking because the cookie name was changed server-side.

Fixes #91.

Signed-off-by: Gergely Nagy <me@gergo.csillger.hu>